+++ server/src/academys/use-cases/findById/find-academy-by-id.spec.ts
@@ -1,6 +1,6 @@
 import { NotFoundException } from '@nestjs/common';
 import { InMemoryAcademysRepository } from '@academys/test/in-memory/InMemoryAcademysRepository';
-import { FindByIdUseCase } from './FindByIdUseCase';
+import { FindByIdUseCase } from './find-academy-by-id';
 
 describe('Find Academy by id', () => {
   let inMemoryAcademysRepository: InMemoryAcademysRepository;
