+++ server/src/academys/use-cases/findById/find-academy-by-id.ts
