+++ server/src/shared/infra/http/controllers/academys.controllers.ts
@@ -1,6 +1,6 @@
 import { Academy } from '@academys/infra/typeorm/entities/academy.entity';
 import { Controller, Get, Param } from '@nestjs/common';
-import { FindByIdUseCase } from './../../../../academys/use-cases/findById/FindByIdUseCase';
+import { FindByIdUseCase } from '../../../../academys/use-cases/findById/find-academy-by-id';
 @Controller('academys')
 export class AcademysControllers {
   constructor(private findByIdUseCase: FindByIdUseCase) {}
